+++ tests/integration/components/preprint-form-project-select-test.js
@@ -13,40 +13,38 @@ moduleForComponent('preprint-form-project-select', 'Integration | Component | pr
 
 test('it renders', function(assert) {
     this.render(hbs`{{preprint-form-project-select
-        existingNodeExistingFile=(action noop)
         changeInitialState=(action noop)
         finishUpload=(action noop)
-        createComponentCopyFile=(action noop)
         highlightSuccessOrFailure=(action noop)
     }}`);
-    assert.equal(this.$('p.text-muted').text().trim(), 'The list of projects appearing in the selector are projects and components for which you have admin access.  Registrations are not included here.');
+    assert.equal(this.$('p.text-muted').text().trim(), 'The list of projects appearing in the selector are projects and components for which you have admin access.');
 });
 
-test('isAdmin computed to false shows warning', function(assert) {
+test('isNodeAdmin computed to false shows warning', function(assert) {
     this.set('selectedNode', {
         currentUserPermissions: [Permissions.ADMIN],
     });
     this.render(hbs`{{preprint-form-project-select
-        existingNodeExistingFile=(action noop)
-        changeInitialState=(action noop)
-        finishUpload=(action noop)
-        createComponentCopyFile=(action noop)
-        selectedNode=selectedNode
-        nodeLocked=true
-        highlightSuccessOrFailure=(action noop)
-    }}`);
+            changeInitialState=(action noop)
+            finishUpload=(action noop)
+            selectedNode=selectedNode
+            isNodeAdmin=true
+            preprintLocked=true
+            currentState='existing'
+            highlightSuccessOrFailure=(action noop)
+        }}`);
     assert.ok(!this.$('.alert-danger').length);
 
     this.set('selectedNode', {
         currentUserPermissions: [],
     });
     this.render(hbs`{{preprint-form-project-select
-        existingNodeExistingFile=(action noop)
         changeInitialState=(action noop)
         finishUpload=(action noop)
-        createComponentCopyFile=(action noop)
         selectedNode=selectedNode
-        nodeLocked=true
+        preprintLocked=true
+        isNodeAdmin=false
+        currentState='existing'
         highlightSuccessOrFailure=(action noop)
     }}`);
     assert.ok(this.$('.alert-danger').length);
@@ -57,10 +55,8 @@ skip('choosing a project locks the node', function() {
     // states in this component, dependant on https://github.com/CenterForOpenScience/ember-preprints/pull/293/files
     test('choosing a project locks the node', function(assert) {
         this.render(hbs`{{preprint-form-project-select
-            existingNodeExistingFile=(action noop)
             changeInitialState=(action noop)
             finishUpload=(action noop)
-            createComponentCopyFile=(action noop)
             userNodesLoaded=true
             userNodes=userNodes
             highlightSuccessOrFailure=(action noop)